Avoiding some more warnings with GCC -Wconversion
git-svn-id: svn://svn.code.sf.net/p/fuego/code/trunk@1190 2e953b5c-c64d-0410-be54-f773e93e544c

+++ go/GoBook.cpp
@@ -221,7 +221,7 @@ const GoBook::MapEntry* GoBook::LookupEntry(const GoBoard& bd) const
 SgPoint GoBook::LookupMove(const GoBoard& bd) const
 {
     vector<SgPoint> moves = LookupAllMoves(bd);
-    int nuMoves = moves.size();
+    size_t nuMoves = moves.size();
     if (nuMoves == 0)
         return SG_NULLMOVE;
     SgPoint p = moves[rand() % nuMoves];
